+++ plugins/woocommerce-admin/client/two-column-tasks/task-list.tsx
@@ -17,7 +17,6 @@ import {
 	TaskType,
 	useUserPreferences,
 	getVisibleTasks,
-	WCDataSelector,
 	TaskListType,
 } from '@woocommerce/data';
 import { recordEvent } from '@woocommerce/tracks';
@@ -58,7 +57,7 @@ export const TaskList: React.FC< TaskListProps > = ( {
 	const { updateOptions, dismissTask, undoDismissTask } = useDispatch(
 		OPTIONS_STORE_NAME
 	);
-	const { profileItems } = useSelect( ( select: WCDataSelector ) => {
+	const { profileItems } = useSelect( ( select ) => {
 		const { getProfileItems } = select( ONBOARDING_STORE_NAME );
 		return {
 			profileItems: getProfileItems(),